requirements.txt
asgiref==3.6.0
astroid==2.14.2
autopep8==2.0.1
bleach==6.0.0
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
dill==0.3.6
Django==4.1.7
django-taggit==3.1.0
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
Markdown==3.3.4
martor==1.6.18
mccabe==0.7.0
Pillow==9.4.0
platformdirs==3.0.0
pycodestyle==2.10.0
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.14
webencodings==0.5.1
wrapt==1.14.1
